Add API tests
