+++ package.json
@@ -44,7 +44,8 @@
     "react-dom": "^15.5.4",
     "react-hot-loader": "3.0.0-beta.6",
     "react-virtualized": "^9.8.0",
-    "rxjs": "^5.4.1"
+    "rxjs": "^5.4.1",
+    "supertest": "^3.0.0"
   },
   "devDependencies": {
     "@webpack-blocks/dev-server2": "^0.4.0",
